LANGCHAIN/pyproject.toml
[project]
name = "langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.37.1",
    "langchain-groq>=1.1.2",
    "langsmith>=0.7.3",
    "python-dotenv>=1.2.1",
]
